fix overtuning on flush thread
also removes more reasons for a large slab to be "dirty".

I was trying to avoid over-flushing when the global page pool is low,
but it looks like the thread doesn't really keep up and we end up not
moving pages fast enough. so the lower the global page pool the more
over-aggressive the flusher gets.

maybe instead of "oldest age" we could sort the large slabs by free
chunks?